+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/IBridge.kt
@@ -26,6 +26,7 @@ internal interface IBridge {
         network: ICaptureNetwork,
         preferences: IPreferences,
         errorReporter: IErrorReporter,
+        issueReportCallback: IssueReportCallback?,
         startInSleepMode: Boolean,
     ): Long
 }